refactor: added results controller and moved results methods to said controller
